refactor: migrate all Vitest test files from src/ to dedicated tests/ directory
