+++ tests/test-utils.tsx
@@ -1,7 +1,7 @@
 import { render, RenderOptions } from '@testing-library/react';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { BrowserRouter } from 'react-router-dom';
-import { ReactElement, ReactNode } from 'react';
+import React, { ReactElement, ReactNode } from 'react';
 
 const createTestQueryClient = () => new QueryClient({
   defaultOptions: {
